tmp/requirements.txt

# Core dependencies
pymongo==4.6.1
python-dotenv==1.0.0
pyyaml==6.0.1

# ML/NLP dependencies
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.2.2
numpy==1.24.3
scipy==1.11.4

# Utilities
tqdm==4.66.1
click==8.1.7
colorama==0.4.6

# Optional: for better performance
accelerate==0.25.0
optimum==1.16.1